+++ androidApp/src/main/java/hr/dtakac/prognoza/presentation/places/PlacesUiMapper.kt
@@ -13,15 +13,34 @@ class PlacesUiMapper @Inject constructor(
     @Named("computation")
     private val computationDispatcher: CoroutineDispatcher
 ) {
-    suspend fun mapToPlaceUi(
+    suspend fun mapToSavedPlacesUi(
         places: List<Place>,
-        selectedPlace: Place?
+        selectedPlace: Place?,
+    ): List<PlaceUi> = mapToPlaceUi(
+        places = places,
+        selectedPlace = selectedPlace,
+        areDeletable = true
+    )
+
+    suspend fun mapToSearchResultPlacesUi(
+        places: List<Place>
+    ): List<PlaceUi> = mapToPlaceUi(
+        places = places,
+        selectedPlace = null,
+        areDeletable = false
+    )
+
+    private suspend fun mapToPlaceUi(
+        places: List<Place>,
+        selectedPlace: Place?,
+        areDeletable: Boolean
     ): List<PlaceUi> = withContext(computationDispatcher) {
         places.map {
             PlaceUi(
                 name = TextResource.fromString(it.name),
                 details = TextResource.fromString(it.details ?: ""),
-                isSelected = it == selectedPlace
+                isSelected = it == selectedPlace,
+                isDeletable = areDeletable && it != selectedPlace
             )
         }
     }