+++ androidApp/src/main/java/hr/dtakac/prognoza/ui/places/PlacesList.kt
@@ -1,6 +1,7 @@
 package hr.dtakac.prognoza.ui.places
 
-import androidx.compose.foundation.clickable
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -8,16 +9,21 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.hapticfeedback.HapticFeedbackType
+import androidx.compose.ui.platform.LocalHapticFeedback
 import androidx.compose.ui.unit.dp
 import hr.dtakac.prognoza.presentation.asString
 import hr.dtakac.prognoza.presentation.places.PlaceUi
 
+@OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun PlacesList(
     places: List<PlaceUi>,
-    onPlaceSelected: (Int) -> Unit,
+    onPlaceClick: (Int) -> Unit,
+    onPlaceLongClick: (Int) -> Unit,
     modifier: Modifier = Modifier
 ) {
+    val hapticFeedback = LocalHapticFeedback.current
     LazyColumn(
         modifier = modifier,
         contentPadding = PaddingValues(top = 12.dp)
@@ -28,7 +34,17 @@ fun PlacesList(
                 details = placeUi.details.asString(),
                 isSelected = placeUi.isSelected,
                 modifier = Modifier
-                    .clickable { onPlaceSelected(idx) }
+                    .combinedClickable(
+                        onClick = { onPlaceClick(idx) },
+                        onLongClick = if (placeUi.isDeletable) {
+                            {
+                                onPlaceLongClick(idx)
+                                hapticFeedback.performHapticFeedback(
+                                    hapticFeedbackType = HapticFeedbackType.LongPress
+                                )
+                            }
+                        } else null
+                    )
                     .fillMaxWidth()
                     .padding(
                         vertical = 12.dp,