+++ androidApp/src/main/java/hr/dtakac/prognoza/ui/places/PlacesScreen.kt
@@ -1,9 +1,15 @@
 package hr.dtakac.prognoza.ui.places
 
+import androidx.compose.material3.Text
 import androidx.compose.runtime.*
+import androidx.compose.ui.platform.LocalFocusManager
+import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
+import hr.dtakac.prognoza.R
 import hr.dtakac.prognoza.presentation.OnEvent
+import hr.dtakac.prognoza.presentation.asString
 import hr.dtakac.prognoza.presentation.places.PlacesViewModel
+import hr.dtakac.prognoza.ui.common.AppDialog
 
 @Composable
 fun PlacesScreen(
@@ -12,23 +18,49 @@ fun PlacesScreen(
     onSettingsClick: () -> Unit = {}
 ) {
     val state by viewModel.state
-    LaunchedEffect(true) {
-        viewModel.getSaved()
-    }
+    var query by remember { mutableStateOf("") }
+    var placeDeletionDialogIndex by remember { mutableStateOf<Int?>(null) }
+    val focusManager = LocalFocusManager.current
+
     OnEvent(state.placeSelected) {
+        query = ""
+        focusManager.clearFocus()
         onPlaceSelected()
     }
 
-    var query by remember { mutableStateOf("") }
     PlacesContent(
         state = state,
-        onPlaceSelected = viewModel::select,
         onSettingsClick = onSettingsClick,
-        query = query,
-        onQuerySubmit = { viewModel.search(query) },
-        onQueryChange = {
-            query = it
-            if (it.isBlank()) viewModel.getSaved()
+        onPlaceClick = { idx ->
+            viewModel.selectPlace(idx)
+        },
+        onPlaceLongClick = { idx ->
+            placeDeletionDialogIndex = idx
+        },
+        searchQuery = query,
+        onSearchSubmit = {
+            viewModel.getSearchResults(query)
+        },
+        onSearchQueryChange = { newQuery ->
+            query = newQuery
+            if (newQuery.isBlank()) {
+                viewModel.getSaved()
+            }
         }
     )
+
+    placeDeletionDialogIndex?.let { idx ->
+        val placeName = state.places[idx].name.asString()
+        AppDialog(
+            title = stringResource(id = R.string.delete_place_title),
+            onConfirm = {
+                viewModel.deletePlace(idx)
+                placeDeletionDialogIndex = null
+            },
+            confirmLabel = stringResource(id = R.string.delete),
+            onDismiss = {
+                placeDeletionDialogIndex = null
+            }
+        ) { Text(stringResource(id = R.string.template_delete_place_description, placeName)) }
+    }
 }