+++ shared/src/commonMain/kotlin/hr/dtakac/prognoza/shared/domain/data/SavedPlaceRemover.kt
@@ -1,0 +1,7 @@
+package hr.dtakac.prognoza.shared.domain.data
+
+import hr.dtakac.prognoza.shared.entity.Place
+
+internal interface SavedPlaceRemover {
+    suspend fun remove(place: Place)
+}