+++ shared/src/iosMain/kotlin/hr/dtakac/prognoza/shared/platform/PrognozaSdkFactory.kt
@@ -1,14 +1,17 @@
 package hr.dtakac.prognoza.shared.platform
 
+import hr.dtakac.prognoza.shared.PrognozaSdk
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.newFixedThreadPoolContext
 
-class IosPrognozaSdkFactory(userAgent: String) : PrognozaSdkFactory(
-    userAgent = userAgent,
-    localRfc2616LanguageGetter = LocalRfc2616LanguageGetter(),
-    sqlDriverFactory = SqlDriverFactory(),
-    dotDecimalFormatter = DotDecimalFormatter(),
-    rfc1123UtcDateTimeParser = Rfc1123UtcDateTimeParser(),
-    ioDispatcher = newFixedThreadPoolContext(nThreads = 64, name = "iOS IO pool"),
-    computationDispatcher = Dispatchers.Default
-)
+actual class PrognozaSdkFactory(private val userAgent: String) {
+    actual fun create(): PrognozaSdk = InternalPrognozaSdkFactory(
+        userAgent = userAgent,
+        localRfc2616LanguageGetter = LocalRfc2616LanguageGetter(),
+        sqlDriverFactory = SqlDriverFactory(),
+        dotDecimalFormatter = DotDecimalFormatter(),
+        rfc1123UtcDateTimeParser = Rfc1123UtcDateTimeParser(),
+        ioDispatcher = newFixedThreadPoolContext(nThreads = 64, name = "iOS IO pool"),
+        computationDispatcher = Dispatchers.Default
+    ).create()
+}