+++ androidApp/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - iOS support
 - NWS weather data provider
 
+## [Unreleased]
+### Added
+- Ability to delete saved places
+
 ## [3.1.0] - 2022-12-09
 ### Added
 - Option to determine mood from device wallpaper on Android 12+